image-processor/requirements.txt
azure-batch==12.0.0
azure-core-tracing-opencensus==1.0.0b8
azure-keyvault==4.2.0
azure-identity==1.11.0
pydantic==1.10.2
python-dotenv==0.21.0
requests==2.28.1
azure-storage-blob==12.14.1
rosbags
cvbridge3
opencv-python